move more paths util functions over
